Extract shared device types and config maps to lib/device.ts
Deduplicate Device type, DeviceType, DEVICE_TYPE_CONFIG, STATUS_CONFIG,
DEVICE_CHANNEL_HINTS, and formatLastSeen across devices.tsx,
device-detail.tsx, and meter-card.tsx into a shared lib/device.ts module.
Add device type validation whitelist in backend CreateParams.Valid().

Closes #46
